feat(suite): implement collapsible legacy types
