+++ packages/components/src/components/CollapsibleBox/CollapsibleBox.tsx
@@ -59,6 +59,7 @@ type CollapsibleBoxCommon = {
     isIconFlipped?: boolean; // Open upwards, affects the icon rotation
     hasDivider?: boolean;
     margin?: FrameProps['margin'];
+    onAnimationComplete?: (isOpen: boolean) => void;
     'data-test'?: string;
     /** @deprecated */
     className?: string;
@@ -198,6 +199,7 @@ const CollapsibleBoxContent = ({
     children,
     margin,
     className,
+    onAnimationComplete,
     'data-test': dataTest,
 }: CollapsibleBoxContentProps) => {
     const { elevation } = useElevation();
@@ -238,6 +240,7 @@ const CollapsibleBoxContent = ({
                 initial={false} // Prevents animation on mount when expanded === false
                 variants={animationVariants}
                 animate={isOpen ? 'expanded' : 'closed'}
+                onAnimationComplete={() => onAnimationComplete?.(isOpen)}
                 transition={{
                     duration: ANIMATION_DURATION,
                     ease: motionEasing.transition,
