+++ packages/suite-analytics/src/types/definitions.ts
@@ -4,7 +4,7 @@ export type OnboardingAnalytics = {
     startTime: number;
     firmware: 'install' | 'update' | 'skip' | 'up-to-date';
     seed: 'create' | 'recovery' | 'recovery-in-progress';
-    seedType: 'shamir-default' | 'shamir-advance' | '12-words' | '24-words';
+    seedType: 'shamir-single' | 'shamir-advanced' | '12-words' | '24-words';
     recoveryType: 'standard' | 'advanced';
     backup: 'create' | 'skip';
     pin: 'create' | 'skip';
