+++ packages/suite-web/e2e/tests/onboarding/t2t1-create-wallet.test.ts
@@ -20,7 +20,7 @@ describe('Onboarding - create wallet', () => {
 
         cy.log('Will be clicking on Shamir backup button');
         cy.getTestElement('@onboarding/select-seed-type-open-dialog').click();
-        cy.getTestElement('@onboarding/select-seed-type-shamir-advance').click();
+        cy.getTestElement('@onboarding/select-seed-type-shamir-advanced').click();
         cy.getTestElement('@onboarding/select-seed-type-confirm').click();
         cy.getTestElement('@onboarding/confirm-on-device').should('be.visible');
         cy.task('pressYes');
